Refactor
- Changed folders to use lowercase
- Fixed typo of commissionRate
- Changed AdressRemover class name
