+++ src/Composition/inversify.config.ts
@@ -2,22 +2,22 @@ import { Container } from "inversify"
 import TYPES from "./Types"
 import "reflect-metadata"
 
-import { IBlockDataProvider, IDataProviderFactory, IStakeDataProvider } from "../Core/DataProvider/Models"
-import { IAddressRemover, IBlockProcessor, IPaymentBuilder, IPaymentProcessor } from '../Core/Payment/Model'
-import { IPayoutCalculator } from "../Core/PayoutCalculator/Model"
-import { BlockProcessor } from "../Core/Payment/BlockProcessor"
-import { PaymentBuilder } from "../Core/Payment/PaymentBuilder"
-import { PaymentProcessor } from "../Core/Payment/PaymentProcessor"
-import { TransactionBuilder } from "../Core/Transaction/TransactionBuilder"
-import { BlockDataProviderFactory } from "../Core/DataProvider/BlockDataProviderFactory"
-import { StakeDataProviderFactory } from "../Core/DataProvider/StakeDataProviderFactory"
-import { IFileWriter } from "../Shared/Model"
-import { FileWriter } from "../Shared/FileWriter"
-import { ISender, ITransactionBuilder, ITransactionProcessor } from "../Core/Transaction/Model"
-import { TransactionSender } from "../Core/Transaction/TrasactionSender"
-import { TransactionProcessor } from "../Core/Transaction/TransactionProcessor"
-import { PayoutCalculatorIsolateSuperCharge } from "../Core/PayoutCalculator/PayoutCalculatorIsolateSuperCharge"
-import { AddressRemoverForSuperCharge } from "../Core/Payment/AddressRemoverForSuperCharge"
+import { IBlockDataProvider, IDataProviderFactory, IStakeDataProvider } from "../core/dataProvider/Models"
+import { IAddressRemover, IBlockProcessor, IPaymentBuilder, IPaymentProcessor } from '../core/payment/Model'
+import { IPayoutCalculator } from "../core/payoutCalculator/Model"
+import { BlockProcessor } from "../core/payment/BlockProcessor"
+import { PaymentBuilder } from "../core/payment/PaymentBuilder"
+import { PaymentProcessor } from "../core/payment/PaymentProcessor"
+import { TransactionBuilder } from "../core/transaction/TransactionBuilder"
+import { BlockDataProviderFactory } from "../core/dataProvider/BlockDataProviderFactory"
+import { StakeDataProviderFactory } from "../core/dataProvider/StakeDataProviderFactory"
+import { IFileWriter } from "../shared/Model"
+import { FileWriter } from "../shared/FileWriter"
+import { ISender, ITransactionBuilder, ITransactionProcessor } from "../core/transaction/Model"
+import { TransactionSender } from "../core/transaction/TrasactionSender"
+import { TransactionProcessor } from "../core/transaction/TransactionProcessor"
+import { PayoutCalculatorIsolateSuperCharge } from "../core/payoutCalculator/PayoutCalculatorIsolateSuperCharge"
+import { AddressRemoverForSuperCharge } from "../core/payment/SubstituteAndExcludePayToAddressesForSuperCharge"
 
 
 var container = new Container()
