+++ src/Tests/block-queries-sql.test2.ts
@@ -1,4 +1,4 @@
-import { getLatestHeight, getBlocks } from '../Core/DataProvider/dataprovider-archivedb/block-queries-sql'
+import { getLatestHeight, getBlocks } from '../core/dataProvider/dataprovider-archivedb/block-queries-sql'
 
 //This tests are not meant to run automatically
 describe('block-queries-sql tests', () => {
