+++ src/core/DataProvider/BlockDataProviderFactory.ts
@@ -3,6 +3,7 @@ import { ArchiveBlockDataProvider } from "./ArchiveBlockDataProvider";
 import { MinaExplorerBlockDataProvider } from "./MinaExplorerBlockDataProvider";
 import { IBlockDataProvider, IDataProviderFactory } from "./Models";
 
+
 @injectable()
 export class BlockDataProviderFactory implements IDataProviderFactory<IBlockDataProvider> {
      build(dataSource: string) : IBlockDataProvider {
