+++ src/core/DataProvider/dataprovider-archivedb/staking-ledger-json-file.ts
@@ -1,6 +1,6 @@
-import { Block, Stake, LedgerEntry } from '../dataprovider-types'
-import { calculateUntimedSlot, getPublicKeyShareClass } from '../../../Utils/staking-ledger-util'
-import fs from "fs";
+
+import { calculateUntimedSlot, getPublicKeyShareClass } from '../../../utils/staking-ledger-util'
+import { LedgerEntry, Stake } from '../dataprovider-types';
 
 // for a given key, find all the stakers delegating to the provided public key (according to the provided epoch staking ledger)
 // determine when key will be unlocked and eligible for supercharged coinbase awards
