+++ src/core/DataProvider/dataprovider-minaexplorer/block-queries-gql.ts
@@ -1,4 +1,4 @@
-import { fetchGraphQL } from '../../../Infrastructure/graphql';
+import { fetchGraphQL } from '../../../infrastructure/graphql';
 import { Blocks } from "../dataprovider-types";
 import fs from "fs";
 import parse from "csv-parse";
